Four catalogue indexes existed twice; the better one now lives at the public URL
/charities, /foundations, /social-enterprises and /grants each had TWO index pages over
identical data: a bespoke public one, and a thin config over the shared browser under
/dashboard/browse. Same rows either way — charities is 66,143 on both.

The shared one wins on every measure. Charities: 86 lines over `OrgBrowser` versus 519
bespoke; the indexed `charity_browse` RPC versus `v_charity_explorer`; sortable columns and
filters as shareable URLs versus neither; and it says why it is empty instead of rendering
blank.

It moves to the PUBLIC url rather than the public url redirecting to it. /dashboard/browse/*
is not where a visitor's entry point should live, and those URLs are the ranked ones.

The shell is a layout, not a place — /search already proved a top-level route can wrap
itself in <Shell>. So the index renders inside the shell AT /charities, and the four paths
join `isChromeless` by EXACT match, not prefix: /charities/[abn], /charities/claim,
/charities/insights and /foundations/minderoo are public detail pages and keep the public
nav. Verified all eight, both ways round.

Shared browsers moved out of app/dashboard/browse/ to components/browse/ — they are no
longer dashboard-specific and leaving them there would have made the import path a lie.

NOT DONE, deliberately: /dashboard/browse/foundations/[id] stays where it is. It shows
grantees, board paths and funding chains, while the public /foundations/[id] shows programs,
opportunities and AIS financials. Those are two different views of one entity, not a
duplicate, and merging them is its own decision. Also untouched: contracts, buyers and
donations have no public twin, so they were not part of this call.

Preserved: the bespoke charities index was the only route into /charities/insights and the
only prompt to claim a listing. Both are links on the new one rather than orphans.
