+++ apps/web/src/app/dashboard/entities/page.tsx
@@ -13,10 +13,10 @@ export const metadata: Metadata = { title: 'Entities — CivicGraph' };
  */
 
 const KINDS: { href: string; label: string; blurb: string }[] = [
-  { href: '/dashboard/browse/foundations', label: 'Foundations', blurb: 'who gives, and to whom' },
-  { href: '/dashboard/browse/social-enterprises', label: 'Social enterprises', blurb: 'the register and what we know' },
-  { href: '/dashboard/browse/charities', label: 'Charities', blurb: 'ACNC register with six years of returns' },
-  { href: '/dashboard/browse/grants', label: 'Grant recipients', blurb: 'who receives justice-system grants' },
+  { href: '/foundations', label: 'Foundations', blurb: 'who gives, and to whom' },
+  { href: '/social-enterprises', label: 'Social enterprises', blurb: 'the register and what we know' },
+  { href: '/charities', label: 'Charities', blurb: 'ACNC register with six years of returns' },
+  { href: '/grants', label: 'Grant recipients', blurb: 'who receives justice-system grants' },
   { href: '/dashboard/browse/contracts', label: 'Contract suppliers', blurb: 'who wins Commonwealth contracts' },
   { href: '/dashboard/browse/buyers', label: 'Government buyers', blurb: 'which agencies let them' },
   { href: '/dashboard/browse/donations', label: 'Political donors', blurb: 'declared donations only' },
